[td-6260]fix compiler error.

+++ src/query/src/qExecutor.c
@@ -6997,7 +6997,7 @@ SOperatorInfo* createSLimitOperatorInfo(SQueryRuntimeEnv* pRuntimeEnv, SOperator
   pInfo->slimit          = pQueryAttr->slimit;
   pInfo->limit           = pQueryAttr->limit;
   pInfo->capacity        = pRuntimeEnv->resultInfo.capacity;
-  pInfo->threshold       = pInfo->capacity * 0.8;
+  pInfo->threshold       = (int64_t) (pInfo->capacity * 0.8);
   pInfo->currentOffset   = pQueryAttr->limit.offset;
   pInfo->currentGroupOffset = pQueryAttr->slimit.offset;
   pInfo->multigroupResult= multigroupResult;
